// src/core/enums/otp.enum.ts
export enum OTP_EVENT {
  OTP_RESET_PASSWORD = 'OTP_RESET_PASSWORD',
  OTP_UPDATE_RESOURCE = 'OTP_UPDATE_RESOURCE',
  OTP_CONFIRMATION_RESOURCE = 'OTP_CONFIRMATION_RESOURCE',
  OTP_CONCILIATION_RESOURCE = 'OTP_CONCILIATION_RESOURCE',
  OTP_INCONSISTENCE_RESOURCE = 'OTP_INCONSISTENCE_RESOURCE',
  OTP_TRANSFERENCE_RESOURCE = 'OTP_TRANSFERENCE_RESOURCE',
  OTP_DELETE_RESOURCE = 'OTP_DELETE_RESOURCE',
  OTP_REMOVE_ITEM_RESOURCE = 'OTP_REMOVE_ITEM_RESOURCE',
  OTP_USER_SIGNIN = 'OTP_USER_SIGNIN',
  OTP_CLOSE_BATCH = 'OTP_CLOSE_BATCH',
  OTP_CREATE_RESOURCE = 'OTP_CREATE_RESOURCE',
  OTP_FINISH_PROCESS_PRODUCTION = 'OTP_FINISH_PROCESS_PRODUCTION',
}